Mineral changing, Edit terrain coordinates, and proper room deletion (#31)
* Add a mineral cycling tool

* Fix the tool coordinate calculations

* Swap middle clicks for alt-clicks

Because I have an Apple trackpad and no way to make a middle click.

* Change the server API to handle proper room deletion

The map-tool will now provide both the new data and the list of rooms it
didn't touch. This allows the server part to check the current map data
against that and do the `map.removeRoom()` call on any rooms it didn't
end up updating.

This means the Generate tool right click can now properly delete rooms.
The old bricking function is still accessible through Alt clicking (both
for rooms and whole sectors).

Also change `makeSolidRoom` to take a room name instead of room coordinates

* Make Show walls not depend on the room generator

Also makes it show the rest of the objects normally instead of skipping

* Add a `getStats` helper